Add `result` parameter to `on_consistent_failure` callback
This is an extension of 42ad14c8da37b220d28b701636c3d0c0d17c264d, which
added the `result` parameter to the `on_failure` and `on_retry`
callbacks. It can be useful to have access to the result in the
`on_consistent_failure` callback, too (and callbacks can still omit the
result parameter if they don't want to use it). Now the callbacks all
support the same parameters again.