Add methods  to retry (#33)
* Add methods_to_retry option

It allows to retry iff the method is on the list.
The list should be populated string with following format:

    methods_to_retry: ["TestClass#testing_something"]

    class TestClass
      def testing_something
      end
    end

When methods_to_retry is empty it retries all.

* Wrong look up on klass.class.name

The 'class' is not necessary there. It was for test, but
now the tests have been updated to behave like a regular class.